Existing file are overwritten if NoResume and too large. Fixes #23
Previously, when Request.NoResume is true, local files
are overwritten if they are smaller than the remote file.
It was assumed that a larger local file meant the files
are not the same and should generate a user-actionable
error.

Now, local files are always overwritten when
Request.NoResume is true, except if the local file
exactly matches the remote file in size.